refactor(mocks): single mocks file, specific routes first
- Merge all MSW handlers into one src/shared/api/mocks.ts
- Register /products/popular before /products/:id (MSW matches the first
  fitting handler — popular products query was 404ing into details route)
- Dropped the dead delay query-param reading from handlers
- Mock types imported relatively from the generated slice (no
  '@/shared/api' imports inside shared/api to avoid import cycles)
