+++ eslint.config.mjs
@@ -103,7 +103,7 @@ export default antfu(
   {
     files: [
       '**/__mocks__/**/*.ts',
-      'src/shared/api/mocks/**/*.ts',
+      'src/shared/api/mocks.ts',
     ],
     rules: {
       'no-restricted-imports': 'off',
