--- src/shared/api/mocks/products/mockProductDtoByIds.ts
@@ -1,8 +1,0 @@
-import type { Product as ProductDto } from '@/shared/api'
-import { productsMock } from '@/shared/lib/server'
-
-export function mockProductDtoByIds(ids: number[]) {
-  return productsMock.filter(product =>
-    ids.includes(product.id),
-  ) as ProductDto[]
-}
