--- src/shared/api/mocks/products/productHandlers.ts
@@ -1,16 +1,0 @@
-import { delay, http, HttpResponse } from 'msw'
-import { env } from '@/shared/lib'
-import { __serverDatabase } from '@/shared/lib/server'
-
-export const productsHandlers = [
-  http.get(`${env.VITE_API_ENDPOINT}/products`, async ({ request }) => {
-    const url = new URL(request.url)
-    const productIds = url.searchParams.getAll('id').map(Number)
-    const products = __serverDatabase.product.findMany(q =>
-      q.where({ id: id => productIds.includes(id) }),
-    )
-
-    await delay(env.VITE_API_DELAY)
-    return HttpResponse.json(products, { status: 200 })
-  }),
-]
