--- src/shared/api/mocks/products/productPopularListHandlers.ts
@@ -1,17 +1,0 @@
-import { delay, http, HttpResponse } from 'msw'
-import { env } from '@/shared/lib'
-import { __serverDatabase } from '@/shared/lib/server'
-
-export const productPopularListHandlers = [
-  http.get(`${env.VITE_API_ENDPOINT}/products/popular`, async () => {
-    const products = __serverDatabase.product.findMany(q =>
-      q.where({ popular: true }),
-    )
-
-    await delay(env.VITE_API_DELAY)
-    return HttpResponse.json(
-      products,
-      { status: 200 },
-    )
-  }),
-]
